Move CarouselSubmissionForm into separate file, rearrange components a bit
